Extract Jshell integration tests to a dedicated trait
