+++ modules/integration/src/test/scala/scala/cli/integration/ReplTests212.scala
@@ -1,3 +1,7 @@
 package scala.cli.integration
 
-class ReplTests212 extends ReplTestDefinitions with ReplAmmoniteTestDefinitions with Test212
+class ReplTests212
+    extends ReplTestDefinitions
+    with ReplAmmoniteTestDefinitions
+    with ReplJShellTestDefinitions
+    with Test212
