+++ modules/integration/src/test/scala/scala/cli/integration/ReplTests213.scala
@@ -1,6 +1,10 @@
 package scala.cli.integration
 
-class ReplTests213 extends ReplTestDefinitions with ReplAmmoniteTestDefinitions with Test213 {
+class ReplTests213
+    extends ReplTestDefinitions
+    with ReplAmmoniteTestDefinitions
+    with ReplJShellTestDefinitions
+    with Test213 {
   for {
     withExplicitScala2SnapshotRepo <- Seq(true, false)
     nightlyVersion      = "2.13.nightly"
